% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Get raw school directory data from NDE}
\usage{
get_raw_directory()
}
\value{
Raw data frame as downloaded from NDE
}
\description{
Downloads the raw school directory Excel file from the Nevada
Department of Education website.
}
\keyword{internal}
