% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from NDE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
Raw data frame from NDE
}
\description{
Downloads enrollment data from the Nevada Department of Education.
Data includes district and building level enrollment by grade and demographics.
}
\keyword{internal}
