% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_intermediate}
\alias{get_raw_enr_intermediate}
\title{Download intermediate format NDE enrollment data (2018-2020)}
\usage{
get_raw_enr_intermediate(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Raw data frame
}
\description{
Downloads enrollment data with consolidated sheets.
Column names are simpler than modern format.
}
\keyword{internal}
