% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{list_enr_years}
\alias{list_enr_years}
\title{List available enrollment data years}
\usage{
list_enr_years()
}
\value{
Integer vector of available years
}
\description{
Returns a vector of years for which enrollment data is available
from the Nevada Department of Education.
}
\details{
Nevada provides enrollment data from 2015-16 (end year 2016) through
2025-26 (end year 2026).
}
\examples{
\dontrun{
available_years <- list_enr_years()
}
}
