% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_district_data}
\alias{process_district_data}
\title{Process district data sheet}
\usage{
process_district_data(df, end_year)
}
\arguments{
\item{df}{Raw data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process district data sheet
}
\keyword{internal}
