% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr_intermediate}
\alias{process_enr_intermediate}
\title{Process intermediate format NDE data (2018-2020)}
\usage{
process_enr_intermediate(raw_list, end_year)
}
\arguments{
\item{raw_list}{Raw data list}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process intermediate format NDE data (2018-2020)
}
\keyword{internal}
