% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr_legacy}
\alias{process_enr_legacy}
\title{Process legacy format NDE data (2016-2017)}
\usage{
process_enr_legacy(raw_list, end_year)
}
\arguments{
\item{raw_list}{Raw data list}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process legacy format NDE data (2016-2017)
}
\keyword{internal}
