% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_school_grades}
\alias{process_school_grades}
\title{Process school grade data sheet}
\usage{
process_school_grades(df, end_year)
}
\arguments{
\item{df}{Raw data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process school grade data sheet
}
\keyword{internal}
