% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{standardize_grade}
\alias{standardize_grade}
\title{Standardize grade level codes}
\usage{
standardize_grade(grades)
}
\arguments{
\item{grades}{Vector of grade levels}
}
\value{
Standardized grade codes
}
\description{
Converts various grade level formats to standard codes.
}
\keyword{internal}
